+++ constants.js
@@ -1,0 +1,20 @@
+const MajorColors = [
+    "white",
+    "red",
+    "black",
+    "yellow",
+    "violet"
+];
+
+const MinorColors = [
+    "blue",
+    "orange",
+    "green",
+    "brown",
+    "slate"
+];
+
+module.exports = {
+    MajorColors : MajorColors,
+    MinorColors : MinorColors
+};
